web_backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-calamine
python-multipart
python-dotenv
supabase
